Move Sp800 stuff to their own files
