+++ src/token.rs
@@ -17,6 +17,7 @@ use super::mechanism;
 use super::object;
 use super::pbkdf2;
 use super::rsa;
+use super::sp800_108;
 use super::storage;
 
 use super::{err_rv, get_random_data, to_rv};
@@ -199,6 +200,7 @@ impl Token {
         hmac::register(&mut token.mechanisms, &mut token.object_factories);
         kdf::register(&mut token.mechanisms, &mut token.object_factories);
         pbkdf2::register(&mut token.mechanisms, &mut token.object_factories);
+        sp800_108::register(&mut token.mechanisms, &mut token.object_factories);
 
         if token.filename.len() > 0 {
             match token.storage.open(&token.filename) {
